feat: Add GRPC enpoints
- use of sqlite engine with sqlc

+++ internal/photo/infrastructure/sqlc/models.go
@@ -12,10 +12,10 @@ type Photo struct {
 	Hash         string         `db:"hash"`
 	Path         string         `db:"path"`
 	DateTime     sql.NullString `db:"date_time"`
-	Iso          sql.NullInt32  `db:"iso"`
+	Iso          sql.NullInt64  `db:"iso"`
 	ExposureTime sql.NullString `db:"exposure_time"`
-	XDimension   sql.NullInt32  `db:"x_dimension"`
-	YDimension   sql.NullInt32  `db:"y_dimension"`
+	XDimension   sql.NullInt64  `db:"x_dimension"`
+	YDimension   sql.NullInt64  `db:"y_dimension"`
 	Model        sql.NullString `db:"model"`
 	FNumber      sql.NullString `db:"f_number"`
 }
